Split monolithic dashboard api.js into focused route modules
The 1,962-line src/dashboard/routes/api.js was a single file containing
all API routes, middleware, security helpers, and validators with no
separation of concerns. This made navigation, testing, and code review
very difficult and caused frequent merge conflicts.

Introduces:
- src/dashboard/lib/middleware.js — shared Express middleware (checkAuth,
  checkGuildAccess, checkWriteRateLimit, checkCsrfOrigin, checkAnyGuildAdmin,
  BoundedRateLimiter)
- src/dashboard/lib/apiHelpers.js — shared utility functions (isValidDiscordId,
  sanitizeMongoValue, logAuditEvent, computeRetention, median,
  parseChannelIdFromJumpUrl)
- src/dashboard/routes/api/ — 15 focused route modules:
  settings, stats, autorole, reactionRoles, rss, knowledgeBase,
  summaryJobs, dailyDigest, ai, members, achievements, itemImages,
  moderation, economy, leveling

api.js is now a 20-line thin orchestrator that mounts all sub-routers.
No behaviour changes — all routes, middleware chains, and validation
logic are preserved exactly.

Closes #89
