+++ src/dashboard/routes/api.js
@@ -1,1962 +1,20 @@
 const express = require('express');
 const router = express.Router();
-const Guild = require('../../models/Guild');
-const Case = require('../../models/Case');
-const User = require('../../models/User');
-const KnowledgeBase = require('../../models/KnowledgeBase');
-const SummaryJob = require('../../models/SummaryJob');
 
-const MAX_SUMMARY_JOBS_PER_GUILD = 10;
-const { rescheduleDailyNews, sendDailyNews } = require('../../services/rssService');
-const { rescheduleBibleVerse } = require('../../services/dailyBibleService');
-const Parser = require('rss-parser');
-const dns = require('dns');
-const net = require('net');
-const http = require('http');
-const https = require('https');
-
-// Returns true for any IP that must not be fetched (loopback, RFC1918, link-local, IPv6 ULA/LL, etc.).
-function isPrivateIp(ip) {
-    if (net.isIPv4(ip)) {
-        const [a, b] = ip.split('.').map(Number);
-        return (
-            a === 0 ||
-            a === 127 ||
-            a === 10 ||
-            (a === 100 && b >= 64 && b <= 127) || // RFC 6598 shared address space
-            (a === 172 && b >= 16 && b <= 31) ||
-            (a === 192 && b === 168) ||
-            (a === 169 && b === 254) ||            // link-local
-            (a >= 224 && a <= 239) ||              // multicast 224.0.0.0/4
-            a >= 240                               // reserved/broadcast 240.0.0.0/4 + 255.255.255.255
-        );
-    }
-    if (net.isIPv6(ip)) {
-        const n = ip.toLowerCase();
-        return (
-            n === '::1' ||                         // loopback
-            n === '::' ||                          // unspecified
-            n.startsWith('fc') ||                  // ULA fc00::/7
-            n.startsWith('fd') ||                  // ULA fd00::/8
-            /^fe[89ab]/i.test(n) ||                // link-local fe80::/10
-            n.startsWith('ff') ||                  // multicast ff00::/8
-            n.startsWith('::ffff:') ||             // IPv4-mapped ::ffff:0:0/96
-            n.startsWith('::ffff:0:') ||           // IPv4-translated (RFC 2765)
-            n.startsWith('64:ff9b:') ||            // IPv4-IPv6 translation (RFC 6052)
-            n.startsWith('2001:db8:') ||           // documentation (RFC 3849)
-            n.startsWith('100::')                  // discard prefix (RFC 6666)
-        );
-    }
-    return true; // unknown format — block by default
-}
-
-// Resolves a hostname to all its IP addresses, validates none are private, and returns
-// the first address to use as a pinned IP for the actual TCP connection.
-// Pinning prevents DNS rebinding: the IP checked here is the IP we connect to.
-async function resolveAndPin(hostname) {
-    const addrs = await new Promise((resolve, reject) => {
-        dns.lookup(hostname, { all: true }, (err, results) => {
-            if (err) reject(new Error(`DNS lookup failed: ${err.message}`));
-            else resolve(results);
-        });
-    });
-    if (!addrs.length) throw new Error('Hostname resolved to no addresses.');
-    for (const { address } of addrs) {
-        if (isPrivateIp(address)) throw new Error('Feed URL resolves to a private or reserved IP address.');
-    }
-    return addrs[0].address; // pinned IP used for the actual connection
-}
-
-// Fetches a feed URL safely: pins DNS on every hop, follows redirects up to maxRedirects.
-// Returns the response body as a string.
-async function safeFetchFeed(urlStr, maxRedirects = 5) {
-    const tls = require('tls');
-    let current = new URL(urlStr);
-
-    for (let hop = 0; hop <= maxRedirects; hop++) {
-        if (!['http:', 'https:'].includes(current.protocol)) {
-            throw new Error('Redirect to non-HTTP protocol rejected.');
-        }
-
-        // Resolve DNS once, validate all returned IPs, then pin to avoid rebinding.
-        const pinnedIp = await resolveAndPin(current.hostname);
-        const port = current.port ? Number(current.port) : (current.protocol === 'https:' ? 443 : 80);
-        const FEED_MAX_BYTES = 5 * 1024 * 1024; // 5 MB — enough for any real RSS feed
-
-        const result = await new Promise((resolve, reject) => {
-            const commonHeaders = {
-                'User-Agent': 'Clawdia-FeedValidator/1.0',
-                Accept: 'application/rss+xml,application/atom+xml,application/xml,text/xml,*/*',
-                Host: current.hostname, // required when connecting directly to a pinned IP
-            };
-
-            let req;
-            if (current.protocol === 'https:') {
-                // For HTTPS: connect to the pinned IP but validate the TLS cert against
-                // the original hostname (SNI + checkServerIdentity).
-                req = https.request({
-                    hostname: current.hostname,
-                    port,
-                    path: current.pathname + current.search,
-                    method: 'GET',
-                    headers: commonHeaders,
-                    timeout: 8000,
-                    createConnection: (opts, cb) => tls.connect({
-                        host: pinnedIp,
-                        port,
-                        servername: current.hostname,
-                        rejectUnauthorized: true,
-                    }, cb),
-                }, handleResponse);
-            } else {
-                req = http.request({
-                    hostname: pinnedIp, // connect to pinned IP directly
-                    port,
-                    path: current.pathname + current.search,
-                    method: 'GET',
-                    headers: commonHeaders,
-                    timeout: 8000,
-                }, handleResponse);
-            }
-
-            function handleResponse(res) {
-                if ([301, 302, 303, 307, 308].includes(res.statusCode)) {
-                    const loc = res.headers.location;
-                    res.destroy();
-                    return resolve({ redirect: loc });
-                }
-                const chunks = [];
-                let totalBytes = 0;
-                res.on('data', c => {
-                    totalBytes += c.length;
-                    if (totalBytes > FEED_MAX_BYTES) {
-                        res.destroy();
-                        return reject(new Error('Feed response exceeds maximum allowed size (5 MB).'));
-                    }
-                    chunks.push(c);
-                });
-                res.on('end', () => resolve({ body: Buffer.concat(chunks).toString('utf8') }));
-                res.on('error', reject);
-            }
-
-            req.on('timeout', () => { req.destroy(); reject(new Error('Feed request timed out.')); });
-            req.on('error', reject);
-            req.end();
-        });
-
-        if (result.redirect) {
-            if (typeof result.redirect !== 'string' || !result.redirect.trim()) {
-                throw new Error('Redirect with empty Location header.');
-            }
-            current = new URL(result.redirect, current.href);
-            continue;
-        }
-        return result.body;
-    }
-    throw new Error('Too many redirects.');
-}
-
-// Filters memberEvents by calendar date rather than array position so sparse
-// histories (days with no events) don't skew the 7/30-day windows.
-function computeRetention(memberEvents, nowMs = Date.now()) {
-    const cutoff7  = new Date(nowMs - 7  * 24 * 60 * 60 * 1000).toISOString().slice(0, 10);
-    const cutoff30 = new Date(nowMs - 30 * 24 * 60 * 60 * 1000).toISOString().slice(0, 10);
-    const events7  = memberEvents.filter(e => e.date >= cutoff7);
-    const events30 = memberEvents.filter(e => e.date >= cutoff30);
-    const joins7   = events7.reduce((a, d) => a + (d.joins  || 0), 0);
-    const leaves7  = events7.reduce((a, d) => a + (d.leaves || 0), 0);
-    const joins30  = events30.reduce((a, d) => a + (d.joins  || 0), 0);
-    const leaves30 = events30.reduce((a, d) => a + (d.leaves || 0), 0);
-    const retained7  = joins7  ? Math.max(0, joins7  - leaves7)  / joins7  : 0;
-    const retained30 = joins30 ? Math.max(0, joins30 - leaves30) / joins30 : 0;
-    return { joins7, leaves7, joins30, leaves30, retained7, retained30 };
-}
-
-function median(nums) {
-    if (!nums.length) return null;
-    const sorted = [...nums].sort((a, b) => a - b);
-    const mid = Math.floor(sorted.length / 2);
-    return sorted.length % 2 ? sorted[mid] : (sorted[mid - 1] + sorted[mid]) / 2;
-}
-
-function parseChannelIdFromJumpUrl(url) {
-    if (!url || typeof url !== 'string') return null;
-    const parts = url.split('/').filter(Boolean);
-    return parts.length >= 2 ? parts[parts.length - 2] : null;
-}
-
-// Discord snowflake IDs are 17–20 digit strings.
-function isValidDiscordId(id) {
-    return typeof id === 'string' && /^\d{17,20}$/.test(id);
-}
-
-function checkAuth(req, res, next) {
-    if (req.isAuthenticated()) return next();
-    res.status(401).json({ error: 'Unauthorized' });
-}
-
-function checkGuildAccess(req, res, next) {
-    const { guildId } = req.params;
-    const userGuilds = req.user.guilds.filter(guild =>
-        (guild.permissions & 0x20) === 0x20 && req.client.guilds.cache.has(guild.id)
-    );
-
-    if (!userGuilds.find(g => g.id === guildId)) {
-        return res.status(403).json({ error: 'Forbidden' });
-    }
-
-    next();
-}
-
-// Bounded in-memory rate limiter.
-// Uses a Map (insertion-order preserved) so that when the map is full the oldest
-// entry is evicted first — giving FIFO/LRU behaviour without a separate library.
-// Max size of 10 000 entries prevents memory exhaustion even under sustained
-// enumeration attacks using unique fake user IDs.
-class BoundedRateLimiter {
-    constructor(maxSize = 10_000) {
-        this._map = new Map();
-        this._maxSize = maxSize;
-    }
-
-    check(userId, windowMs, limit) {
-        const now = Date.now();
-        const arr = (this._map.get(userId) || []).filter(t => now - t < windowMs);
-        if (arr.length >= limit) {
-            this._map.set(userId, arr); // refresh pruned array
-            return false;
-        }
-        // Evict oldest entry before inserting a new key
-        if (!this._map.has(userId) && this._map.size >= this._maxSize) {
-            this._map.delete(this._map.keys().next().value);
-        }
-        arr.push(now);
-        this._map.set(userId, arr);
-        return true;
-    }
-
-    cleanup(windowMs) {
-        const cutoff = Date.now() - windowMs;
-        for (const [userId, timestamps] of this._map) {
-            if (timestamps.every(t => t < cutoff)) this._map.delete(userId);
-        }
-    }
-}
-
-const WRITE_RL_WINDOW_MS = 60 * 1000;
-const WRITE_RL_LIMIT = 60;
-const writeRateLimiter = new BoundedRateLimiter(10_000);
-// Clean up stale entries every minute (was every 5 min — tightened to reduce memory growth window)
-setInterval(() => writeRateLimiter.cleanup(WRITE_RL_WINDOW_MS), 60 * 1000).unref();
-
-function checkWriteRateLimit(req, res, next) {
-    const userId = req.user?.id;
-    if (!userId) return res.status(401).json({ error: 'Unauthorized' });
-    if (!writeRateLimiter.check(userId, WRITE_RL_WINDOW_MS, WRITE_RL_LIMIT)) {
-        return res.status(429).json({ error: 'Too many requests. Please slow down.' });
-    }
-    next();
-}
-
-// M2: CSRF origin validation for all state-changing API requests.
-// Complements sameSite: 'lax' cookies — rejects cross-origin POST/PUT/DELETE that
-// carry an Origin header pointing to a different host than the dashboard.
-function checkCsrfOrigin(req, res, next) {
-    const origin = req.headers.origin;
-    if (!origin) return next(); // same-origin requests may omit Origin
-    const dashboardUrl = process.env.DASHBOARD_URL || `http://localhost:${process.env.DASHBOARD_PORT || 3000}`;
-    try {
-        if (new URL(origin).origin === new URL(dashboardUrl).origin) return next();
-    } catch { /* fall through to reject */ }
-    return res.status(403).json({ error: 'Forbidden: cross-origin request rejected' });
-}
-
-// H1: Recursively strip any object key starting with '$' to prevent NoSQL operator
-// injection. Mongoose schema type validation is the primary defence; this is a
-// belt-and-suspenders layer applied before any .set() call.
-function sanitizeMongoValue(value) {
-    if (value === null || value === undefined || typeof value !== 'object') return value;
-    if (Array.isArray(value)) return value.map(sanitizeMongoValue);
-    const out = {};
-    for (const [k, v] of Object.entries(value)) {
-        if (k.startsWith('$')) continue;
-        out[k] = sanitizeMongoValue(v);
-    }
-    return out;
-}
-
-// L1: Structured audit log — writes to the AuditLog collection. Failures are
-// swallowed so audit errors never block the main operation.
-async function logAuditEvent(req, guildId, action, details = null) {
-    try {
-        const AuditLog = require('../../models/AuditLog');
-        await AuditLog.create({
-            guildId,
-            userId:    req.user?.id || 'unknown',
-            action,
-            ip:        req.ip || null,
-            userAgent: req.get('user-agent') || null,
-            details,
-        });
-    } catch (err) {
-        console.error('[AUDIT] Failed to write audit event:', err.message);
-    }
-}
-
-// Top-level Guild schema keys that the dashboard is allowed to update.
-// This whitelist prevents prototype pollution (__proto__, constructor, etc.)
-// and limits surface area to only fields the dashboard UI actually manages.
-// Keep in sync with the field names the saveSettings() function in
-// guild-settings.ejs actually sends — anything missing causes a 400.
-const ALLOWED_SETTING_PARENTS = new Set([
-    'welcome', 'farewell', 'birthdays',
-    'moderation', 'leveling', 'levelRoles',
-    'economy', 'shop', 'jobs', 'jobTiers',
-    'achievements',
-    'raidDetection', 'antiNuke', 'caseSettings',
-    'starboard', 'eventLog', 'quests',
-    'season', 'progressionTracks', 'commandPolicies',
-    'suggestions', 'ai', 'tempVoice', 'bibleVerse',
-    'dailyNews', 'dailyNewsProfiles', 'rssFeeds',
-    'autoRoles', 'reactionRoles', 'analytics',
-    'giveaways', 'notifications',
-    'newspaper', 'heist', 'exploration'
-]);
-
-function isAllowedSettingKey(key) {
-    if (typeof key !== 'string') return false;
-    const top = key.split('.')[0];
-    return ALLOWED_SETTING_PARENTS.has(top);
-}
-
-// Field-level validation for welcome settings before they reach Mongoose (fix #12).
-// Returns an error string, or null if valid.
-function validateWelcomeUpdate(updates) {
-    for (const [key, value] of Object.entries(updates)) {
-        if (!key.startsWith('welcome.') && key !== 'welcome') continue;
-        const field = key.split('.')[1];
-        if (field === 'message' || field === 'dmMessage') {
-            if (typeof value !== 'string') return `welcome.${field} must be a string`;
-            if (value.length > 4000) return `welcome.${field} exceeds 4000 characters`;
-        }
-        if (field === 'enabled' || field === 'cardEnabled' || field === 'dmEnabled') {
-            if (typeof value !== 'boolean') return `welcome.${field} must be a boolean`;
-        }
-        if (field === 'channelId' && value !== null && value !== '') {
-            if (typeof value !== 'string' || !/^\d{17,20}$/.test(value)) {
-                return 'welcome.channelId must be a valid Discord snowflake or null';
-            }
-        }
-    }
-    return null;
-}
-
-function validateFarewellUpdate(updates) {
-    for (const [key, value] of Object.entries(updates)) {
-        if (!key.startsWith('farewell.') && key !== 'farewell') continue;
-        const field = key.split('.')[1];
-        if (field === 'message') {
-            if (typeof value !== 'string') return 'farewell.message must be a string';
-            if (value.length > 4000) return 'farewell.message exceeds 4000 characters';
-        }
-        if (field === 'enabled') {
-            if (typeof value !== 'boolean') return 'farewell.enabled must be a boolean';
-        }
-        if (field === 'channelId' && value !== null && value !== '') {
-            if (typeof value !== 'string' || !/^\d{17,20}$/.test(value)) {
-                return 'farewell.channelId must be a valid Discord snowflake or null';
-            }
-        }
-    }
-    return null;
-}
-
-function validateBirthdaysUpdate(updates) {
-    for (const [key, value] of Object.entries(updates)) {
-        if (!key.startsWith('birthdays.') && key !== 'birthdays') continue;
-        const field = key.split('.')[1];
-        if (field === 'message') {
-            if (typeof value !== 'string') return 'birthdays.message must be a string';
-            if (value.length > 2000) return 'birthdays.message exceeds 2000 characters';
-        }
-        if (field === 'enabled') {
-            if (typeof value !== 'boolean') return 'birthdays.enabled must be a boolean';
-        }
-        if (field === 'channelId' && value !== null && value !== '') {
-            if (typeof value !== 'string' || !/^\d{17,20}$/.test(value)) {
-                return 'birthdays.channelId must be a valid Discord snowflake or null';
-            }
-        }
-        if (field === 'roleId' && value !== null && value !== '') {
-            if (typeof value !== 'string' || !/^\d{17,20}$/.test(value)) {
-                return 'birthdays.roleId must be a valid Discord snowflake or null';
-            }
-        }
-        if (field === 'wishingHourUtc') {
-            if (typeof value !== 'number' || !Number.isInteger(value) || value < 0 || value > 23) {
-                return 'birthdays.wishingHourUtc must be an integer between 0 and 23';
-            }
-        }
-    }
-    return null;
-}
-
-function validateEventLogUpdate(updates) {
-    const EVENT_LOG_BOOLEAN_FIELDS = new Set([
-        'enabled', 'logMessageEdit', 'logMessageDelete', 'logMessageBulkDelete',
-        'logMemberJoin', 'logMemberLeave', 'logNicknameChange', 'logUsernameChange',
-        'logAvatarChange', 'logTimeout', 'logBoost',
-        'logRoleChanges', 'logChannelChanges',
-        'logVoiceJoin', 'logVoiceLeave', 'logVoiceMove', 'logVoiceMuteDeafen',
-        'logInviteCreate', 'logInviteDelete', 'logServerUpdate', 'logEmojiUpdate',
-        'logWebhookUpdate', 'logBotAdd',
-        'logThreadCreate', 'logThreadDelete', 'logThreadArchive',
-    ]);
-    for (const [key, value] of Object.entries(updates)) {
-        if (!key.startsWith('eventLog.') && key !== 'eventLog') continue;
-        const field = key.split('.')[1];
-        if (EVENT_LOG_BOOLEAN_FIELDS.has(field)) {
-            if (typeof value !== 'boolean') return `eventLog.${field} must be a boolean`;
-        }
-        if (field === 'channelId' && value !== null && value !== '') {
-            if (typeof value !== 'string' || !/^\d{17,20}$/.test(value)) {
-                return 'eventLog.channelId must be a valid Discord snowflake or null';
-            }
-        }
-    }
-    return null;
-}
-
-const VALID_BIBLE_TRANSLATIONS = new Set(['kjv', 'niv', 'asv', 'web', 'ylt', 'darby', 'bbe', 'webbe']);
-
-function validateBibleVerseUpdate(updates) {
-    for (const [key, value] of Object.entries(updates)) {
-        if (!key.startsWith('bibleVerse.') && key !== 'bibleVerse') continue;
-        const field = key.split('.')[1];
-        if (field === 'enabled' || field === 'autoRespond') {
-            if (typeof value !== 'boolean') return `bibleVerse.${field} must be a boolean`;
-        }
-        if (field === 'channelId' && value !== null && value !== '') {
-            if (typeof value !== 'string' || !/^\d{17,20}$/.test(value)) {
-                return 'bibleVerse.channelId must be a valid Discord snowflake or null';
-            }
-        }
-        if (field === 'time' && value !== null && value !== '') {
-            if (typeof value !== 'string' || !/^\d{1,2}:\d{2}$/.test(value)) {
-                return 'bibleVerse.time must be in HH:MM format (e.g. 08:00)';
-            }
-            const [h, m] = value.split(':').map(Number);
-            if (h < 0 || h > 23 || m < 0 || m > 59) {
-                return 'bibleVerse.time must be a valid 24-hour time (00:00–23:59)';
-            }
-        }
-        if (field === 'timezone' && value !== null && value !== '') {
-            if (typeof value !== 'string') return 'bibleVerse.timezone must be a string';
-            try {
-                Intl.DateTimeFormat(undefined, { timeZone: value });
-            } catch {
-                return `bibleVerse.timezone "${value}" is not a valid IANA timezone`;
-            }
-        }
-        if (field === 'translation') {
-            if (!VALID_BIBLE_TRANSLATIONS.has(value)) {
-                return `bibleVerse.translation must be one of: ${[...VALID_BIBLE_TRANSLATIONS].join(', ')}`;
-            }
-        }
-    }
-    return null;
-}
-
-function validateSnowflakeOrNull(value, label) {
-    if (value === null || value === '' || value === undefined) return null;
-    if (typeof value !== 'string' || !/^\d{17,20}$/.test(value)) {
-        return `${label} must be a valid Discord snowflake or null`;
-    }
-    return null;
-}
-
-function validateNewspaperUpdate(updates) {
-    for (const [key, value] of Object.entries(updates)) {
-        if (!key.startsWith('newspaper.') && key !== 'newspaper') continue;
-        const field = key.split('.')[1];
-        if (field === 'channelId') {
-            const err = validateSnowflakeOrNull(value, 'newspaper.channelId');
-            if (err) return err;
-        }
-        if (field === 'quoteChannelIds') {
-            if (!Array.isArray(value) || value.length > 100) return 'newspaper.quoteChannelIds must be an array of at most 100 channel ids';
-            for (const id of value) {
-                const err = validateSnowflakeOrNull(id, 'newspaper.quoteChannelIds entry');
-                if (err) return err;
-            }
-        }
-    }
-    return null;
-}
-
-// Field-level validation for exploration settings before they reach Mongoose.
-// Returns an error string, or null if valid.
-function validateExplorationUpdate(updates) {
-    for (const [key, value] of Object.entries(updates)) {
-        if (!key.startsWith('exploration.') && key !== 'exploration') continue;
-        const field = key.split('.')[1];
-        if (field === 'enabled' || field === 'announceSecrets') {
-            if (typeof value !== 'boolean') return `exploration.${field} must be a boolean`;
-        }
-        if (field === 'dropRateMultiplier') {
-            if (typeof value !== 'number' || !Number.isFinite(value) || value < 0.1 || value > 5) {
-                return 'exploration.dropRateMultiplier must be a number between 0.1 and 5';
-            }
-        }
-        if (field === 'rareEventBonus') {
-            if (typeof value !== 'number' || !Number.isFinite(value) || value < 0 || value > 0.25) {
-                return 'exploration.rareEventBonus must be a number between 0 and 0.25';
-            }
-        }
-        if (field === 'disabledRegions') {
-            if (!Array.isArray(value) || value.length > 100 || value.some(v => typeof v !== 'string' || v.length > 64)) {
-                return 'exploration.disabledRegions must be an array of at most 100 region id strings';
-            }
-        }
-    }
-    return null;
-}
-
-function validateHeistUpdate(updates) {
-    for (const [key, value] of Object.entries(updates)) {
-        if (!key.startsWith('heist.') && key !== 'heist') continue;
-        const field = key.split('.')[1];
-        if (field === 'announceChannelId') {
-            const err = validateSnowflakeOrNull(value, 'heist.announceChannelId');
-            if (err) return err;
-        }
-    }
-    return null;
-}
-
-router.post('/guild/:guildId/settings', checkAuth, checkGuildAccess, checkCsrfOrigin, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const updates = req.body;
-
-    if (!updates || typeof updates !== 'object' || Array.isArray(updates)) {
-        return res.status(400).json({ error: 'Request body must be a plain object' });
-    }
-
-    const rejectedKeys = Object.keys(updates).filter(k => !isAllowedSettingKey(k));
-    if (rejectedKeys.length) {
-        return res.status(400).json({ error: `Disallowed setting key(s): ${rejectedKeys.join(', ')}` });
-    }
-
-    const welcomeError = validateWelcomeUpdate(updates);
-    if (welcomeError) return res.status(400).json({ error: welcomeError });
-
-    const farewellError = validateFarewellUpdate(updates);
-    if (farewellError) return res.status(400).json({ error: farewellError });
-
-    const birthdaysError = validateBirthdaysUpdate(updates);
-    if (birthdaysError) return res.status(400).json({ error: birthdaysError });
-
-    const bibleVerseError = validateBibleVerseUpdate(updates);
-    if (bibleVerseError) return res.status(400).json({ error: bibleVerseError });
-
-    const eventLogError = validateEventLogUpdate(updates);
-    if (eventLogError) return res.status(400).json({ error: eventLogError });
-
-    const newspaperError = validateNewspaperUpdate(updates);
-    if (newspaperError) return res.status(400).json({ error: newspaperError });
-
-    const heistError = validateHeistUpdate(updates);
-    if (heistError) return res.status(400).json({ error: heistError });
-
-    const explorationError = validateExplorationUpdate(updates);
-    if (explorationError) return res.status(400).json({ error: explorationError });
-
-    try {
-        const guildSettings = await Guild.findOne({ guildId });
-
-        if (!guildSettings) {
-            return res.status(404).json({ error: 'Guild not found' });
-        }
-
-        // Shop item images live inline at guild.shop[].imageData. The dashboard
-        // never round-trips those Buffers, so a naive replace of the shop array
-        // would wipe every saved image. Preserve imageData/imageType from the
-        // existing items, matched by itemId.
-        if (Array.isArray(updates.shop)) {
-            const existingImages = new Map();
-            for (const item of guildSettings.shop || []) {
-                if (item && item.itemId && item.imageData) {
-                    existingImages.set(item.itemId, { imageData: item.imageData, imageType: item.imageType });
-                }
-            }
-            updates.shop = updates.shop.map(item => {
-                if (!item || !item.itemId) return item;
-                const preserved = existingImages.get(item.itemId);
-                if (!preserved) return item;
-                return { ...item, imageData: preserved.imageData, imageType: preserved.imageType };
-            });
-        }
-
-        // H1: Strip any MongoDB operator keys ($ne, $regex, etc.) before writing.
-        Object.keys(updates).forEach(key => {
-            guildSettings.set(key, sanitizeMongoValue(updates[key]));
-        });
-
-        await guildSettings.save();
-
-        // L1: Record what changed and who changed it.
-        await logAuditEvent(req, guildId, 'settings_update', { keys: Object.keys(updates) });
-
-        const shouldRescheduleDailyNews = Object.keys(updates).some(key => key.startsWith('dailyNews.') || key === 'dailyNewsProfiles');
-        if (shouldRescheduleDailyNews) {
-            rescheduleDailyNews(req.client, guildId);
-        }
-
-        const shouldRescheduleBible = Object.keys(updates).some(key => key.startsWith('bibleVerse.'));
-        if (shouldRescheduleBible) {
-            rescheduleBibleVerse(req.client, guildId);
-        }
-
-        res.json({ success: true, settings: guildSettings });
-    } catch (error) {
-        if (error.name === 'ValidationError') {
-            return res.status(400).json({ error: error.message });
-        }
-        console.error('API error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.get('/guild/:guildId/stats', checkAuth, checkGuildAccess, async (req, res) => {
-    const { guildId } = req.params;
-
-    try {
-        const totalUsers = await User.countDocuments({ guildId });
-        const totalMessages = await User.aggregate([
-            { $match: { guildId } },
-            { $group: { _id: null, total: { $sum: '$messages' } } }
-        ]);
-
-        const topLevels = await User.find({ guildId })
-            .sort({ level: -1, xp: -1 })
-            .limit(10);
-        const guildSettings = await Guild.findOne({ guildId });
-        const memberEvents = guildSettings?.analytics?.memberEvents || [];
-        const commandUsage = guildSettings?.analytics?.commandUsage || [];
-
-        const { joins7, leaves7, joins30, leaves30, retained7, retained30 } = computeRetention(memberEvents);
-
-        const commandSummary = {};
-        const failedByReason = {};
-        const bestTimesByChannel = {};
-        for (const item of commandUsage) {
-            commandSummary[item.command] = commandSummary[item.command] || { total: 0, failed: 0 };
-            commandSummary[item.command].total += 1;
-            if (!item.success) {
-                commandSummary[item.command].failed += 1;
-                failedByReason[item.reason || 'unknown'] = (failedByReason[item.reason || 'unknown'] || 0) + 1;
-            }
-            const channel = item.channelId || 'unknown';
-            bestTimesByChannel[channel] = bestTimesByChannel[channel] || {};
-            bestTimesByChannel[channel][item.hour] = (bestTimesByChannel[channel][item.hour] || 0) + 1;
-        }
-        const bestPostingTimes = Object.entries(bestTimesByChannel).map(([channelId, hours]) => {
-            const bestHour = Object.entries(hours).sort((a, b) => b[1] - a[1])[0]?.[0] ?? '0';
-            return { channelId, hourUtc: Number(bestHour) };
-        }).slice(0, 8);
-
-        const recommendations = [];
-        if (leaves30 > joins30 * 0.6 && !guildSettings?.welcome?.dmEnabled) recommendations.push('Enable welcome DMs to improve first-week retention.');
-        if (!guildSettings?.moderation?.enabled || !guildSettings?.moderation?.autoModEnabled) recommendations.push('Enable auto-moderation, your raid/spam risk is elevated.');
-        if ((failedByReason.execution_error || 0) > 10) recommendations.push('High command error volume detected. Audit recent command updates.');
-        if ((guildSettings?.rssFeeds?.length || 0) === 0) recommendations.push('Add RSS or Daily News automation to keep channels active.');
-
-        // Member growth: derive from memberEvents (last 30 days)
-        const now30 = Date.now();
-        const memberGrowth = [];
-        for (let i = 29; i >= 0; i--) {
-            const d = new Date(now30 - i * 864e5).toISOString().slice(0, 10);
-            const ev = memberEvents.find(e => e.date === d);
-            memberGrowth.push({ date: d, joins: ev?.joins || 0, leaves: ev?.leaves || 0 });
-        }
-
-        // Message volume proxy: commandUsage events grouped by day (last 30 days)
-        const msgVolMap = {};
-        for (const ev of commandUsage) {
-            if (ev.createdAt) {
-                const d = new Date(ev.createdAt).toISOString().slice(0, 10);
-                msgVolMap[d] = (msgVolMap[d] || 0) + 1;
-            }
-        }
-        const messageVolume = memberGrowth.map(({ date }) => ({ date, count: msgVolMap[date] || 0 }));
-
-        // Economy stats summary
-        const [ecoTotalAgg, ecoActiveCount] = await Promise.all([
-            User.aggregate([{ $match: { guildId } }, { $group: { _id: null, total: { $sum: { $add: ['$balance', '$bank'] } }, avgXp: { $avg: '$xp' } } }]),
-            User.countDocuments({ guildId, $or: [
-                { lastWork:  { $gte: new Date(now30 - 7 * 864e5) } },
-                { lastDaily: { $gte: new Date(now30 - 7 * 864e5) } },
-                { lastFish:  { $gte: new Date(now30 - 7 * 864e5) } },
-                { lastMine:  { $gte: new Date(now30 - 7 * 864e5) } },
-                { lastCrime: { $gte: new Date(now30 - 7 * 864e5) } },
-                { lastHeist: { $gte: new Date(now30 - 7 * 864e5) } },
-                { lastRob:   { $gte: new Date(now30 - 7 * 864e5) } }
-            ] })
-        ]);
-
-        res.json({
-            totalUsers,
-            totalMessages: totalMessages[0]?.total || 0,
-            topLevels: topLevels.map(u => ({
-                userId: u.userId,
-                level: u.level,
-                xp: u.xp
-            })),
-            analytics: {
-                growthFunnel: { joins30, retained7: Number((retained7 * 100).toFixed(1)), retained30: Number((retained30 * 100).toFixed(1)) },
-                churnAlerts: leaves30 > joins30 * 0.5 ? ['Churn is elevated over the last 30 days.'] : [],
-                likelyCauses: [
-                    !guildSettings?.welcome?.enabled ? 'Welcome flow is disabled.' : null,
-                    !guildSettings?.leveling?.enabled ? 'No progression loop (leveling disabled).' : null,
-                    !guildSettings?.economy?.enabled ? 'No recurring incentive loop (economy disabled).' : null
-                ].filter(Boolean),
-                bestPostingTimes,
-                commandUsage: commandSummary,
-                failedCommands: failedByReason,
-                recommendations,
-                messageVolume,
-                memberGrowth,
-                economyStats: {
-                    totalCoins: ecoTotalAgg[0]?.total || 0,
-                    activeUsers: ecoActiveCount
-                },
-                xpStats: {
-                    avgXp: ecoTotalAgg[0]?.avgXp ? Math.round(ecoTotalAgg[0].avgXp) : 0
-                }
-            }
-        });
-    } catch (error) {
-        console.error('Stats error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.get('/guild/:guildId/insights', checkAuth, checkGuildAccess, async (req, res) => {
-    const { guildId } = req.params;
-
-    try {
-        const guildSettings = await Guild.findOne({ guildId });
-        if (!guildSettings) return res.status(404).json({ error: 'Guild not found' });
-
-        const memberEvents = guildSettings?.analytics?.memberEvents || [];
-        const commandUsage = guildSettings?.analytics?.commandUsage || [];
-
-        // Retention: 7/30 day net-retention proxy from join/leave tracking.
-        const { joins7, leaves7, joins30, leaves30, retained7, retained30 } = computeRetention(memberEvents);
-
-        // Active hours: command-driven activity histogram (UTC).
-        const hourMap = Array.from({ length: 24 }, (_, hour) => ({ hourUtc: hour, count: 0 }));
-        for (const event of commandUsage) {
-            if (typeof event.hour === 'number' && event.hour >= 0 && event.hour <= 23) {
-                hourMap[event.hour].count += 1;
-            }
-        }
-        const topActiveHours = [...hourMap].sort((a, b) => b.count - a.count).slice(0, 5);
-
-        // Toxic channel hotspot proxy from moderation case evidence jump URLs.
-        const recentCases = await Case.find({ guildId }).sort({ createdAt: -1 }).limit(1000);
-        const channelToxicity = new Map();
-        for (const c of recentCases) {
-            const channelId = parseChannelIdFromJumpUrl(c?.evidence?.jumpUrl) || 'unknown';
-            const current = channelToxicity.get(channelId) || { channelId, incidents: 0, warns: 0, severe: 0, score: 0 };
-            current.incidents += 1;
-            if (c.type === 'warn') current.warns += 1;
-            if (['mute', 'kick', 'ban'].includes(c.type)) current.severe += 1;
-            current.score = current.warns + (current.severe * 2);
-            channelToxicity.set(channelId, current);
-        }
-        const toxicChannels = [...channelToxicity.values()].sort((a, b) => b.score - a.score).slice(0, 8);
-
-        // Moderator SLA: median time to close case + trend grouped by month.
-        const resolvedCases = recentCases.filter(c => c.createdAt && c.resolvedAt);
-        const resolutionHours = resolvedCases.map(c => (new Date(c.resolvedAt) - new Date(c.createdAt)) / 36e5).filter(h => h >= 0);
-        const medianResolutionHours = median(resolutionHours);
-
-        const monthlyTrend = {};
-        for (const c of resolvedCases) {
-            const dt = new Date(c.resolvedAt);
-            const key = `${dt.getUTCFullYear()}-${String(dt.getUTCMonth() + 1).padStart(2, '0')}`;
-            if (!monthlyTrend[key]) monthlyTrend[key] = [];
-            monthlyTrend[key].push((new Date(c.resolvedAt) - new Date(c.createdAt)) / 36e5);
-        }
-        const modSlaTrends = Object.entries(monthlyTrend)
-            .map(([month, arr]) => ({ month, medianResolutionHours: Number((median(arr) || 0).toFixed(2)), resolvedCases: arr.length }))
-            .sort((a, b) => a.month.localeCompare(b.month))
-            .slice(-6);
-
-        // Newcomer conversion after 7/30 days based on user activity.
-        const now = Date.now();
-        const users = await User.find({ guildId }).select('createdAt messages level');
-        const cohort7 = users.filter(u => u.createdAt && (now - new Date(u.createdAt).getTime()) >= 7 * 864e5);
-        const cohort30 = users.filter(u => u.createdAt && (now - new Date(u.createdAt).getTime()) >= 30 * 864e5);
-        const isConverted = (u) => (u.messages || 0) >= 20 || (u.level || 0) >= 2;
-        const converted7 = cohort7.filter(isConverted).length;
-        const converted30 = cohort30.filter(isConverted).length;
-
-        res.json({
-            retention: {
-                joins7,
-                leaves7,
-                retained7Pct: Number((retained7 * 100).toFixed(1)),
-                joins30,
-                leaves30,
-                retained30Pct: Number((retained30 * 100).toFixed(1))
-            },
-            activeHours: {
-                timezone: 'UTC',
-                histogram: hourMap,
-                topHours: topActiveHours
-            },
-            toxicChannels,
-            modSla: {
-                medianResolutionHours: medianResolutionHours == null ? null : Number(medianResolutionHours.toFixed(2)),
-                trends: modSlaTrends
-            },
-            newcomerConversion: {
-                definition: 'Converted = at least 20 messages or level 2+',
-                days7: { cohortSize: cohort7.length, converted: converted7, pct: cohort7.length ? Number(((converted7 / cohort7.length) * 100).toFixed(1)) : 0 },
-                days30: { cohortSize: cohort30.length, converted: converted30, pct: cohort30.length ? Number(((converted30 / cohort30.length) * 100).toFixed(1)) : 0 }
-            }
-        });
-    } catch (error) {
-        console.error('Insights error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.post('/guild/:guildId/autorole', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { roleId } = req.body;
-
-    if (!roleId) return res.status(400).json({ error: 'roleId required' });
-    if (!isValidDiscordId(roleId)) return res.status(400).json({ error: 'roleId must be a valid Discord snowflake' });
-
-    try {
-        const guildSettings = await Guild.findOne({ guildId });
-        if (!guildSettings) return res.status(404).json({ error: 'Guild not found' });
-
-        if (!guildSettings.autoRoles.some(r => r.roleId === roleId)) {
-            guildSettings.autoRoles.push({ roleId });
-            await guildSettings.save();
-        }
-
-        res.json({ success: true });
-    } catch (error) {
-        console.error('Autorole add error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.delete('/guild/:guildId/autorole/:roleId', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId, roleId } = req.params;
-
-    try {
-        const guildSettings = await Guild.findOne({ guildId });
-        if (!guildSettings) return res.status(404).json({ error: 'Guild not found' });
-
-        guildSettings.autoRoles = guildSettings.autoRoles.filter(r => r.roleId !== roleId);
-        await guildSettings.save();
-
-        res.json({ success: true });
-    } catch (error) {
-        console.error('Autorole remove error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.post('/guild/:guildId/reactionrole/panel', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { channelId, title, description, mappings } = req.body;
-
-    if (!channelId) return res.status(400).json({ error: 'channelId is required' });
-    if (!Array.isArray(mappings) || !mappings.length) return res.status(400).json({ error: 'At least one emoji/role mapping is required' });
-
-    for (const m of mappings) {
-        if (!m || typeof m.emoji !== 'string' || !m.emoji.trim() ||
-            typeof m.roleId !== 'string' || !m.roleId.trim()) {
-            return res.status(400).json({ error: 'Each mapping must have a non-empty emoji and roleId' });
-        }
-        if (!isValidDiscordId(m.roleId.trim())) {
-            return res.status(400).json({ error: `Invalid roleId: ${m.roleId} — must be a valid Discord snowflake` });
-        }
-    }
-
-    const emojiValues = mappings.map(m => m.emoji.trim());
-    if (new Set(emojiValues).size !== emojiValues.length) {
-        return res.status(400).json({ error: 'Duplicate emoji values are not allowed within the same panel' });
-    }
-
-    try {
-        const guild = req.client.guilds.cache.get(guildId);
-        if (!guild) return res.status(404).json({ error: 'Guild not found' });
-
-        const channel = guild.channels.cache.get(channelId);
-        if (!channel) return res.status(404).json({ error: 'Channel not found' });
-
-        const guildSettings = await Guild.findOne({ guildId });
-        if (!guildSettings) return res.status(404).json({ error: 'Guild settings not found' });
-
-        const { EmbedBuilder } = require('discord.js');
-        const embed = new EmbedBuilder()
-            .setColor(0x5865F2)
-            .setTitle(title || 'React to get a role!')
-            .setDescription(
-                (description ? description + '\n\n' : '') +
-                mappings.map(m => `${m.emoji.trim()} — <@&${m.roleId.trim()}>`).join('\n')
-            )
-            .setFooter({ text: 'React below to assign yourself a role' });
-
-        const message = await channel.send({ embeds: [embed] });
-
-        try {
-            for (const mapping of mappings) {
-                const emojiStr = mapping.emoji.trim();
-                const match = emojiStr.match(/^<a?:(\w+):(\d+)>$/);
-                const reactArg = match ? `${match[1]}:${match[2]}` : emojiStr;
-                await message.react(reactArg);
-            }
-
-            for (const mapping of mappings) {
-                guildSettings.reactionRoles.push({
-                    messageId: message.id,
-                    channelId,
-                    emoji: mapping.emoji.trim(),
-                    roleId: mapping.roleId.trim()
-                });
-            }
-
-            await guildSettings.save();
-        } catch (innerError) {
-            await message.delete().catch(() => null);
-            throw innerError;
-        }
-
-        res.json({ success: true, messageId: message.id });
-    } catch (error) {
-        console.error('Reaction role panel create error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.delete('/guild/:guildId/reactionrole/panel/:messageId', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId, messageId } = req.params;
-
-    try {
-        const guildSettings = await Guild.findOne({ guildId });
-        if (!guildSettings) return res.status(404).json({ error: 'Guild not found' });
-
-        const entry = guildSettings.reactionRoles.find(r => r.messageId === messageId);
-        if (entry) {
-            const guild = req.client.guilds.cache.get(guildId);
-            if (guild) {
-                const channel = guild.channels.cache.get(entry.channelId);
-                if (channel) {
-                    await channel.messages.fetch(messageId).then(m => m.delete()).catch(() => null);
-                }
-            }
-        }
-
-        guildSettings.reactionRoles = guildSettings.reactionRoles.filter(r => r.messageId !== messageId);
-        await guildSettings.save();
-
-        res.json({ success: true });
-    } catch (error) {
-        console.error('Reaction role panel delete error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.post('/guild/:guildId/validate-feed', checkAuth, checkGuildAccess, async (req, res) => {
-    const { url } = req.body;
-    if (!url || typeof url !== 'string') {
-        return res.status(400).json({ valid: false, error: 'No URL provided.' });
-    }
-
-    let parsed;
-    try {
-        parsed = new URL(url);
-    } catch {
-        return res.json({ valid: false, error: 'Invalid URL format.' });
-    }
-
-    if (!['http:', 'https:'].includes(parsed.protocol)) {
-        return res.json({ valid: false, error: 'URL must use http or https.' });
-    }
-
-    try {
-        const body = await safeFetchFeed(url);
-        const feedParser = new Parser();
-        const feed = await feedParser.parseString(body);
-        return res.json({ valid: true, title: feed.title || '', itemCount: feed.items?.length ?? 0 });
-    } catch (err) {
-        return res.json({ valid: false, error: err.message || 'Could not fetch or parse feed. Check the URL and ensure it is a valid RSS/Atom feed.' });
-    }
-});
-
-router.post('/guild/:guildId/rss/add', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { url, channelId } = req.body;
-
-    if (!url || typeof url !== 'string') return res.status(400).json({ error: 'url is required' });
-    if (!channelId || !isValidDiscordId(channelId)) return res.status(400).json({ error: 'channelId must be a valid Discord snowflake' });
-
-    let parsedUrl;
-    try {
-        parsedUrl = new URL(url);
-    } catch {
-        return res.status(400).json({ error: 'url must be a valid URL' });
-    }
-    if (!['http:', 'https:'].includes(parsedUrl.protocol)) {
-        return res.status(400).json({ error: 'url must use http or https' });
-    }
-
-    try {
-        const guildSettings = await Guild.findOne({ guildId });
-        if (!guildSettings) return res.status(404).json({ error: 'Guild not found' });
-
-        guildSettings.rssFeeds.push({ url: url.trim(), channelId });
-        await guildSettings.save();
-
-        res.json({ success: true });
-    } catch (error) {
-        console.error('RSS add error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-const dailyNewsInFlight = new Set();
-router.post('/guild/:guildId/dailynews/trigger', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    if (dailyNewsInFlight.has(guildId)) {
-        return res.status(409).json({ error: 'A digest is already being sent for this guild. Please wait for it to finish.' });
-    }
-    dailyNewsInFlight.add(guildId);
-    try {
-        await sendDailyNews(req.client, guildId);
-        res.json({ success: true });
-    } catch (error) {
-        console.error('Daily news manual trigger error:', error);
-        res.status(500).json({ error: 'Failed to send daily news. Check that the digest is configured.' });
-    } finally {
-        dailyNewsInFlight.delete(guildId);
-    }
-});
-
-router.delete('/guild/:guildId/rss/:index', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId, index } = req.params;
-
-    try {
-        const guildSettings = await Guild.findOne({ guildId });
-        
-        guildSettings.rssFeeds.splice(parseInt(index), 1);
-        await guildSettings.save();
-
-        res.json({ success: true });
-    } catch (error) {
-        console.error('RSS delete error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.get('/guild/:guildId/knowledge-base', checkAuth, checkGuildAccess, async (req, res) => {
-    const { guildId } = req.params;
-    try {
-        const entries = await KnowledgeBase.find({ guildId }).sort({ createdAt: -1 }).limit(100);
-        res.json(entries);
-    } catch (error) {
-        console.error('Knowledge base list error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.post('/guild/:guildId/knowledge-base', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { title, content, tags } = req.body;
-
-    if (!title || typeof title !== 'string' || !title.trim()) {
-        return res.status(400).json({ error: 'Title is required' });
-    }
-    if (!content || typeof content !== 'string' || !content.trim()) {
-        return res.status(400).json({ error: 'Content is required' });
-    }
-
-    const sanitizedTags = Array.isArray(tags) ? tags.map(t => String(t).trim()).filter(Boolean).slice(0, 10) : [];
-
-    try {
-        const entry = await KnowledgeBase.create({
-            guildId,
-            title: title.trim().slice(0, 200),
-            content: content.trim().slice(0, 4000),
-            tags: sanitizedTags,
-            addedBy: req.user.id
-        });
-        res.json({ success: true, entry });
-    } catch (error) {
-        console.error('Knowledge base add error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.delete('/guild/:guildId/knowledge-base/:entryId', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId, entryId } = req.params;
-
-    if (!/^[0-9a-f]{24}$/i.test(entryId)) {
-        return res.status(400).json({ error: 'Invalid entry ID' });
-    }
-
-    try {
-        const result = await KnowledgeBase.deleteOne({ _id: entryId, guildId });
-        if (result.deletedCount === 0) {
-            return res.status(404).json({ error: 'Entry not found' });
-        }
-        res.json({ success: true });
-    } catch (error) {
-        console.error('Knowledge base delete error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.put('/guild/:guildId/knowledge-base/:entryId', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId, entryId } = req.params;
-    const { title, content, tags } = req.body;
-
-    if (!/^[0-9a-f]{24}$/i.test(entryId)) {
-        return res.status(400).json({ error: 'Invalid entry ID' });
-    }
-    if (!title || typeof title !== 'string' || !title.trim()) {
-        return res.status(400).json({ error: 'Title is required' });
-    }
-    if (!content || typeof content !== 'string' || !content.trim()) {
-        return res.status(400).json({ error: 'Content is required' });
-    }
-
-    const sanitizedTags = Array.isArray(tags) ? tags.map(t => String(t).trim()).filter(Boolean).slice(0, 10) : [];
-
-    try {
-        const entry = await KnowledgeBase.findOneAndUpdate(
-            { _id: entryId, guildId },
-            {
-                title: title.trim().slice(0, 200),
-                content: content.trim().slice(0, 4000),
-                tags: sanitizedTags
-            },
-            { new: true }
-        );
-        if (!entry) {
-            return res.status(404).json({ error: 'Entry not found' });
-        }
-        res.json({ success: true, entry });
-    } catch (error) {
-        console.error('Knowledge base update error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// ── AI Summary Jobs ────────────────────────────────────────────────────────
-
-router.get('/guild/:guildId/summary-jobs', checkAuth, checkGuildAccess, async (req, res) => {
-    const { guildId } = req.params;
-    try {
-        const jobs = await SummaryJob.find({ guildId }).sort({ createdAt: 1 });
-        res.json(jobs);
-    } catch (error) {
-        console.error('Summary jobs list error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.post('/guild/:guildId/summary-jobs', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { sourceChannelId, targetChannelId, hour, minute, label } = req.body;
-
-    if (!sourceChannelId || typeof sourceChannelId !== 'string' || !sourceChannelId.trim()) {
-        return res.status(400).json({ error: 'sourceChannelId is required' });
-    }
-    if (!targetChannelId || typeof targetChannelId !== 'string' || !targetChannelId.trim()) {
-        return res.status(400).json({ error: 'targetChannelId is required' });
-    }
-    const h = parseInt(hour, 10);
-    const m = parseInt(minute, 10);
-    if (!Number.isFinite(h) || h < 0 || h > 23) return res.status(400).json({ error: 'hour must be 0–23' });
-    if (!Number.isFinite(m) || m < 0 || m > 59) return res.status(400).json({ error: 'minute must be 0–59' });
-
-    try {
-        const guild = req.client.guilds.cache.get(guildId);
-        if (!guild) return res.status(404).json({ error: 'Guild not found' });
-
-        const srcChannel = guild.channels.cache.get(sourceChannelId.trim());
-        const tgtChannel = guild.channels.cache.get(targetChannelId.trim());
-        if (!srcChannel || !tgtChannel) return res.status(400).json({ error: 'One or both channels not found in this guild' });
-
-        const jobCount = await SummaryJob.countDocuments({ guildId });
-        if (jobCount >= MAX_SUMMARY_JOBS_PER_GUILD) {
-            return res.status(400).json({ error: `Maximum of ${MAX_SUMMARY_JOBS_PER_GUILD} summary jobs per server reached.` });
-        }
-
-        const jobLabel = (typeof label === 'string' && label.trim())
-            ? label.trim().slice(0, 100)
-            : `Daily Summary of #${srcChannel.name}`.slice(0, 100);
-
-        const job = await SummaryJob.create({
-            guildId,
-            sourceChannelId: sourceChannelId.trim(),
-            targetChannelId: targetChannelId.trim(),
-            hour: h,
-            minute: m,
-            label: jobLabel
-        });
-        res.json({ success: true, job });
-    } catch (error) {
-        console.error('Summary job create error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.delete('/guild/:guildId/summary-jobs/:jobId', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId, jobId } = req.params;
-
-    if (!/^[0-9a-f]{24}$/i.test(jobId)) {
-        return res.status(400).json({ error: 'Invalid job ID' });
-    }
-
-    try {
-        const result = await SummaryJob.deleteOne({ _id: jobId, guildId });
-        if (result.deletedCount === 0) return res.status(404).json({ error: 'Job not found' });
-        res.json({ success: true });
-    } catch (error) {
-        console.error('Summary job delete error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// ── Daily Digest ───────────────────────────────────────────────────────────
-
-router.get('/guild/:guildId/daily-digest', checkAuth, checkGuildAccess, async (req, res) => {
-    const { guildId } = req.params;
-    try {
-        const settings = await Guild.findOne({ guildId });
-        if (!settings) return res.status(404).json({ error: 'Guild not found' });
-        res.json(settings.ai?.dailyDigest || {});
-    } catch (error) {
-        console.error('Daily digest get error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-const SNOWFLAKE_RE = /^\d{17,19}$/;
-function isSnowflake(id) { return typeof id === 'string' && SNOWFLAKE_RE.test(id); }
-
-router.put('/guild/:guildId/daily-digest', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { enabled, channelId, sourceChannelIds, hour, minute, timezone } = req.body;
-
-    if (hour !== undefined && !/^\d+$/.test(String(hour))) return res.status(400).json({ error: 'hour must be a non-negative integer' });
-    if (minute !== undefined && !/^\d+$/.test(String(minute))) return res.status(400).json({ error: 'minute must be a non-negative integer' });
-    const h = hour !== undefined ? parseInt(hour, 10) : undefined;
-    const m = minute !== undefined ? parseInt(minute, 10) : undefined;
-    if (h !== undefined && (h < 0 || h > 23)) return res.status(400).json({ error: 'hour must be 0–23' });
-    if (m !== undefined && (m < 0 || m > 59)) return res.status(400).json({ error: 'minute must be 0–59' });
-
-    try {
-        const settings = await Guild.findOne({ guildId });
-        if (!settings) return res.status(404).json({ error: 'Guild not found' });
-
-        if (!settings.ai) settings.ai = {};
-        if (!settings.ai.dailyDigest) settings.ai.dailyDigest = {};
-
-        if (typeof enabled === 'boolean') settings.ai.dailyDigest.enabled = enabled;
-        if (channelId !== undefined) settings.ai.dailyDigest.channelId = isSnowflake(channelId) ? channelId : null;
-        if (Array.isArray(sourceChannelIds)) {
-            settings.ai.dailyDigest.sourceChannelIds = [...new Set(sourceChannelIds.filter(isSnowflake))];
-        }
-        if (h !== undefined) settings.ai.dailyDigest.hour = h;
-        if (m !== undefined) settings.ai.dailyDigest.minute = m;
-        if (typeof timezone === 'string' && timezone.trim()) {
-            try {
-                Intl.DateTimeFormat(undefined, { timeZone: timezone.trim() });
-                settings.ai.dailyDigest.timezone = timezone.trim();
-            } catch {
-                return res.status(400).json({ error: 'Invalid timezone' });
-            }
-        }
-
-        settings.markModified('ai');
-        await settings.save();
-        res.json({ success: true, dailyDigest: settings.ai.dailyDigest });
-    } catch (error) {
-        console.error('Daily digest update error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// ── AI Personas ────────────────────────────────────────────────────────────
-
-router.post('/guild/:guildId/persona', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { channelId, personaName, systemPrompt } = req.body;
-
-    if (!channelId || typeof channelId !== 'string' || !channelId.trim()) {
-        return res.status(400).json({ error: 'channelId is required' });
-    }
-    if (!personaName || typeof personaName !== 'string' || !personaName.trim()) {
-        return res.status(400).json({ error: 'personaName is required' });
-    }
-    if (!systemPrompt || typeof systemPrompt !== 'string' || !systemPrompt.trim()) {
-        return res.status(400).json({ error: 'systemPrompt is required' });
-    }
-
-    try {
-        const guild = req.client.guilds.cache.get(guildId);
-        if (!guild) return res.status(404).json({ error: 'Guild not found' });
-
-        const channel = guild.channels.cache.get(channelId.trim());
-        if (!channel) return res.status(400).json({ error: 'Channel not found in this guild' });
-
-        const cid = channelId.trim();
-        const pName = personaName.trim().slice(0, 100);
-        const pPrompt = systemPrompt.trim().slice(0, 4000);
-
-        // Try to update an existing persona for this channel atomically.
-        let result = await Guild.findOneAndUpdate(
-            { guildId, 'ai.channelPersonas.channelId': cid },
-            { $set: { 'ai.channelPersonas.$.personaName': pName, 'ai.channelPersonas.$.systemPrompt': pPrompt } },
-            { new: true }
-        );
-
-        // No matching persona — push a new one.
-        if (!result) {
-            result = await Guild.findOneAndUpdate(
-                { guildId },
-                { $push: { 'ai.channelPersonas': { channelId: cid, personaName: pName, systemPrompt: pPrompt } } },
-                { new: true }
-            );
-        }
-
-        if (!result) return res.status(404).json({ error: 'Guild settings not found' });
-        res.json({ success: true, personas: result.ai?.channelPersonas || [] });
-    } catch (error) {
-        console.error('Persona set error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.delete('/guild/:guildId/persona/:channelId', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId, channelId } = req.params;
-
-    try {
-        const result = await Guild.findOneAndUpdate(
-            { guildId, 'ai.channelPersonas.channelId': channelId },
-            { $pull: { 'ai.channelPersonas': { channelId } } }
-        );
-        if (!result) return res.status(404).json({ error: 'Persona not found for that channel' });
-        res.json({ success: true });
-    } catch (error) {
-        console.error('Persona remove error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// ── AI Usage ───────────────────────────────────────────────────────────────
-
-router.get('/guild/:guildId/ai/usage', checkAuth, checkGuildAccess, async (req, res) => {
-    const { guildId } = req.params;
-    const days = Math.min(90, Math.max(1, parseInt(req.query.days, 10) || 14));
-
-    try {
-        const { getUsageStats } = require('../../services/aiService');
-        const guildSettings = await Guild.findOne({ guildId }).lean();
-        const stats = await getUsageStats(guildId, days);
-
-        const ai = guildSettings?.ai || {};
-        res.json({
-            ...stats,
-            rateLimit: {
-                perUser: ai.rateLimitPerUser ?? 0,
-                perChannel: ai.rateLimitPerChannel ?? 0,
-                windowMin: ai.rateLimitWindowMin ?? 10
-            }
-        });
-    } catch (error) {
-        console.error('AI usage stats error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// ── Member Search ────────────────────────────────────────────────────────────
-
-router.get('/guild/:guildId/members/search', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const q = (req.query.q || '').trim();
-    if (q.length < 2) return res.json([]);
-    try {
-        const guild = req.client.guilds.cache.get(guildId);
-        if (!guild) return res.status(404).json({ error: 'Guild not found' });
-        const results = await guild.members.search({ query: q, limit: 10 });
-        res.json(results.map(m => ({
-            id: m.user.id,
-            username: m.user.username,
-            displayName: m.displayName,
-            avatarURL: m.user.displayAvatarURL({ size: 32, extension: 'webp' })
-        })));
-    } catch (err) {
-        console.error('Member search error:', err);
-        res.status(500).json({ error: 'Search failed' });
-    }
-});
-
-router.get('/guild/:guildId/members/resolve', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const ids = (req.query.ids || '').split(',').map(s => s.trim()).filter(s => /^\d{17,20}$/.test(s)).slice(0, 50);
-    if (!ids.length) return res.json({});
-    try {
-        const result = {};
-        await Promise.all(ids.map(async id => {
-            try {
-                const user = await req.client.users.fetch(id, { force: false });
-                result[id] = { id, username: user.username, displayName: user.globalName || user.username, avatarURL: user.displayAvatarURL({ size: 32, extension: 'webp' }) };
-            } catch { result[id] = null; }
-        }));
-        res.json(result);
-    } catch (err) {
-        console.error('Member resolve error:', err);
-        res.status(500).json({ error: 'Resolve failed' });
-    }
-});
-
-// ── Achievements ────────────────────────────────────────────────────────────
-
-router.post('/guild/:guildId/achievements/grant', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { userId, achievementId } = req.body;
-
-    if (!userId || typeof userId !== 'string' || !userId.trim()) {
-        return res.status(400).json({ error: 'userId is required' });
-    }
-    if (!achievementId || typeof achievementId !== 'string' || !achievementId.trim()) {
-        return res.status(400).json({ error: 'achievementId is required' });
-    }
-
-    try {
-        const guildSettings = await Guild.findOne({ guildId });
-        if (!guildSettings) return res.status(404).json({ error: 'Guild not found' });
-
-        const trimmedId = achievementId.trim();
-        const validAchievement = (guildSettings.achievements?.customAchievements || [])
-            .find(a => a.id === trimmedId);
-        if (!validAchievement) {
-            return res.status(404).json({ error: 'Achievement not found in this server\'s custom achievements' });
-        }
-
-        const User = require('../../models/User');
-        const { grantCustomAchievement } = require('../../services/achievementService');
-
-        let user = await User.findOne({ userId: userId.trim(), guildId });
-        if (!user) {
-            user = await User.create({ userId: userId.trim(), guildId });
-        }
-
-        const granted = await grantCustomAchievement(user, trimmedId);
-        user.markModified('achievements');
-        await user.save();
-
-        res.json({ success: true, granted });
-    } catch (error) {
-        console.error('Achievement grant error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// ── Item Images ────────────────────────────────────────────────────────────────
-
-const multer = require('multer');
-const ItemImage = require('../../models/ItemImage');
-
-// M4: Validate image files by magic bytes rather than trusting the client-supplied
-// MIME type. Prevents disguised file uploads (e.g. PHP named as image/jpeg).
-const IMAGE_SIGNATURES = [
-    { mime: 'image/jpeg', offset: 0, bytes: [0xFF, 0xD8, 0xFF] },
-    { mime: 'image/png',  offset: 0, bytes: [0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A] },
-    { mime: 'image/gif',  offset: 0, bytes: [0x47, 0x49, 0x46, 0x38] },
-    { mime: 'image/webp', offset: 8, bytes: [0x57, 0x45, 0x42, 0x50] },
-];
-
-function detectImageType(buffer) {
-    for (const sig of IMAGE_SIGNATURES) {
-        if (buffer.length < sig.offset + sig.bytes.length) continue;
-        if (sig.bytes.every((b, i) => buffer[sig.offset + i] === b)) return sig.mime;
-    }
-    return null;
-}
-
-const _uploadRaw = multer({
-    storage: multer.memoryStorage(),
-    limits: { fileSize: 512 * 1024 },
-    fileFilter: (_req, file, cb) => {
-        if (!file.mimetype.startsWith('image/')) return cb(new Error('Only image files are allowed'));
-        cb(null, true);
-    }
-});
-
-function uploadImage(req, res, next) {
-    _uploadRaw.single('image')(req, res, err => {
-        if (!err) return next();
-        const status = (err instanceof multer.MulterError || err.message === 'Only image files are allowed') ? 400 : 500;
-        res.status(status).json({ error: err.message || 'Upload error' });
-    });
-}
-
-function checkAnyGuildAdmin(req, res, next) {
-    if (!req.isAuthenticated()) return res.status(401).json({ error: 'Unauthorized' });
-    const adminGuilds = req.user.guilds.filter(g => (g.permissions & 0x20) === 0x20 && req.client.guilds.cache.has(g.id));
-    if (!adminGuilds.length) return res.status(403).json({ error: 'Forbidden' });
-    next();
-}
-
-// Serve a guild shop item's image
-router.get('/item-image/shop/:guildId/:itemId', async (req, res) => {
-    try {
-        const guild = await require('../../models/Guild').findOne({ guildId: req.params.guildId }, { shop: 1 });
-        const item = guild?.shop?.find(i => i.itemId === req.params.itemId);
-        if (!item?.imageData?.length) return res.status(404).end();
-        res.set('Content-Type', item.imageType || 'image/png');
-        res.set('Cache-Control', 'public, max-age=86400');
-        res.send(item.imageData);
-    } catch { res.status(500).end(); }
-});
-
-// Upload image for a guild shop item
-router.post('/item-image/shop/:guildId/:itemId', checkAuth, checkGuildAccess, checkWriteRateLimit, uploadImage, async (req, res) => {
-    if (!req.file) return res.status(400).json({ error: 'No image file provided' });
-    // M4: Verify file contents match a known image signature.
-    const detectedType = detectImageType(req.file.buffer);
-    if (!detectedType) return res.status(400).json({ error: 'Invalid image file: unrecognized format' });
-    try {
-        const guild = await require('../../models/Guild').findOne({ guildId: req.params.guildId });
-        if (!guild) return res.status(404).json({ error: 'Guild not found' });
-        const item = guild.shop.find(i => i.itemId === req.params.itemId);
-        if (!item) return res.status(404).json({ error: 'Shop item not found' });
-        item.imageData = req.file.buffer;
-        item.imageType = detectedType; // use detected type, not client-supplied MIME
-        await guild.save();
-        res.json({ success: true });
-    } catch (err) {
-        console.error('Shop item image upload error:', err);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// Remove image from a guild shop item
-router.delete('/item-image/shop/:guildId/:itemId', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    try {
-        const guild = await require('../../models/Guild').findOne({ guildId: req.params.guildId });
-        if (!guild) return res.status(404).json({ error: 'Guild not found' });
-        const item = guild.shop.find(i => i.itemId === req.params.itemId);
-        if (!item) return res.status(404).json({ error: 'Shop item not found' });
-        item.imageData = null;
-        await guild.save();
-        res.json({ success: true });
-    } catch (err) {
-        console.error('Shop item image delete error:', err);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// Serve a global activity item image (hunt/fish/mine)
-router.get('/item-image/activity/:itemId', async (req, res) => {
-    try {
-        const img = await ItemImage.findOne({ itemId: req.params.itemId });
-        if (!img?.imageData?.length) return res.status(404).end();
-        res.set('Content-Type', img.imageType || 'image/png');
-        res.set('Cache-Control', 'public, max-age=86400');
-        res.send(img.imageData);
-    } catch { res.status(500).end(); }
-});
-
-// Upload/replace a global activity item image (any guild admin)
-router.post('/item-image/activity/:itemId', checkAuth, checkAnyGuildAdmin, checkWriteRateLimit, uploadImage, async (req, res) => {
-    if (!req.file) return res.status(400).json({ error: 'No image file provided' });
-    const { itemId } = req.params;
-    if (!/^[a-z0-9_:\-]{1,64}$/.test(itemId)) return res.status(400).json({ error: 'Invalid itemId' });
-    // M4: Verify file contents match a known image signature.
-    const detectedType = detectImageType(req.file.buffer);
-    if (!detectedType) return res.status(400).json({ error: 'Invalid image file: unrecognized format' });
-    try {
-        await ItemImage.findOneAndUpdate(
-            { itemId },
-            { imageData: req.file.buffer, imageType: detectedType, updatedAt: new Date() },
-            { upsert: true }
-        );
-        res.json({ success: true });
-    } catch (err) {
-        console.error('Activity item image upload error:', err);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// Remove a global activity item image
-router.delete('/item-image/activity/:itemId', checkAuth, checkAnyGuildAdmin, checkWriteRateLimit, async (req, res) => {
-    try {
-        await ItemImage.deleteOne({ itemId: req.params.itemId });
-        res.json({ success: true });
-    } catch (err) {
-        console.error('Activity item image delete error:', err);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// ── Moderation Cases ───────────────────────────────────────────────────────
-
-router.get('/guild/:guildId/cases', checkAuth, checkGuildAccess, async (req, res) => {
-    const { guildId } = req.params;
-    const page = Math.max(1, parseInt(req.query.page, 10) || 1);
-    const limit = Math.min(50, Math.max(1, parseInt(req.query.limit, 10) || 20));
-    const type = req.query.type || null;
-    const status = req.query.status || null;
-
-    try {
-        const query = { guildId };
-        if (type && ['warn', 'mute', 'kick', 'ban', 'unban', 'unmute', 'note', 'appeal'].includes(type)) query.type = type;
-        if (status && ['open', 'closed', 'appealed', 'appeal_approved', 'appeal_denied'].includes(status)) query.status = status;
-
-        const [cases, total] = await Promise.all([
-            Case.find(query).sort({ createdAt: -1 }).skip((page - 1) * limit).limit(limit).lean(),
-            Case.countDocuments(query)
-        ]);
-
-        const uniqueIds = [...new Set(cases.flatMap(c => [c.targetUserId, c.moderatorId].filter(Boolean)))];
-        const userMap = {};
-        await Promise.all(uniqueIds.map(async id => {
-            try {
-                const u = await req.client.users.fetch(id, { force: false });
-                userMap[id] = { tag: u.tag, avatarUrl: u.displayAvatarURL({ size: 32, extension: 'webp' }) };
-            } catch { /* user not resolvable */ }
-        }));
-
-        res.json({
-            cases: cases.map(c => ({
-                ...c,
-                targetUserTag: userMap[c.targetUserId]?.tag || null,
-                targetAvatarUrl: userMap[c.targetUserId]?.avatarUrl || null,
-                moderatorTag: userMap[c.moderatorId]?.tag || null,
-            })),
-            total, page, pages: Math.ceil(total / limit)
-        });
-    } catch (error) {
-        console.error('Cases list error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.patch('/guild/:guildId/cases/:caseId', checkAuth, checkGuildAccess, checkCsrfOrigin, checkWriteRateLimit, async (req, res) => {
-    const { guildId, caseId } = req.params;
-    const { action, note, resolution } = req.body;
-
-    if (!action || !['add_note', 'close'].includes(action)) {
-        return res.status(400).json({ error: 'action must be "add_note" or "close"' });
-    }
-
-    const parsedId = parseInt(caseId, 10);
-    if (!Number.isFinite(parsedId)) return res.status(400).json({ error: 'Invalid caseId' });
-
-    try {
-        const c = await Case.findOne({ guildId, caseId: parsedId });
-        if (!c) return res.status(404).json({ error: 'Case not found' });
-
-        if (action === 'add_note') {
-            if (!note || typeof note !== 'string' || !note.trim()) {
-                return res.status(400).json({ error: 'note is required for add_note' });
-            }
-            c.notes.push({ moderatorId: req.user.id, content: note.trim().slice(0, 1000) });
-        } else if (action === 'close') {
-            c.status = 'closed';
-            c.resolvedAt = new Date();
-            c.resolvedBy = req.user.id;
-            if (resolution && typeof resolution === 'string') {
-                c.resolution = resolution.trim().slice(0, 500);
-            }
-        }
-
-        await c.save();
-        await logAuditEvent(req, guildId, 'case_update', { caseId: parsedId, action });
-        res.json({ success: true, case: c });
-    } catch (error) {
-        console.error('Case update error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.get('/guild/:guildId/sanctions/active', checkAuth, checkGuildAccess, async (req, res) => {
-    const { guildId } = req.params;
-    try {
-        const guild = req.client.guilds.cache.get(guildId);
-        if (!guild) return res.status(404).json({ error: 'Guild not found or bot not in guild' });
-
-        let bans;
-        try {
-            bans = await guild.bans.fetch({ limit: 200 });
-        } catch (banErr) {
-            console.error('Active sanctions: bans fetch failed:', banErr);
-            return res.status(503).json({ error: 'Could not fetch bans from Discord. Check bot permissions.' });
-        }
-
-        const now = new Date();
-        const timeoutMembers = [...guild.members.cache.values()]
-            .filter(m => m.communicationDisabledUntil && m.communicationDisabledUntil > now)
-            .slice(0, 200);
-
-        const banList = [...bans.values()].map(b => ({
-            type: 'ban',
-            userId: b.user.id,
-            userTag: b.user.tag,
-            avatarUrl: b.user.displayAvatarURL({ size: 32 }),
-            reason: b.reason || null,
-            expires: null
-        }));
-
-        const timeoutList = timeoutMembers.map(m => ({
-            type: 'timeout',
-            userId: m.user.id,
-            userTag: m.user.tag,
-            avatarUrl: m.user.displayAvatarURL({ size: 32 }),
-            reason: null,
-            expires: m.communicationDisabledUntil?.toISOString() || null
-        }));
-
-        res.json({ bans: banList, timeouts: timeoutList });
-    } catch (error) {
-        console.error('Active sanctions error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.post('/guild/:guildId/sanctions/unban/:userId', checkAuth, checkGuildAccess, checkCsrfOrigin, checkWriteRateLimit, async (req, res) => {
-    const { guildId, userId } = req.params;
-    if (!isValidDiscordId(userId)) return res.status(400).json({ error: 'Invalid userId' });
-    try {
-        const guild = req.client.guilds.cache.get(guildId);
-        if (!guild) return res.status(404).json({ error: 'Guild not found' });
-        await guild.members.unban(userId, `Unbanned via dashboard by ${req.user.username}`);
-        await logAuditEvent(req, guildId, 'unban', { targetUserId: userId });
-        res.json({ success: true });
-    } catch (error) {
-        console.error('Unban error:', error);
-        res.status(500).json({ error: error.message || 'Internal server error' });
-    }
-});
-
-router.post('/guild/:guildId/sanctions/untimeout/:userId', checkAuth, checkGuildAccess, checkCsrfOrigin, checkWriteRateLimit, async (req, res) => {
-    const { guildId, userId } = req.params;
-    if (!isValidDiscordId(userId)) return res.status(400).json({ error: 'Invalid userId' });
-    try {
-        const guild = req.client.guilds.cache.get(guildId);
-        if (!guild) return res.status(404).json({ error: 'Guild not found' });
-        const member = await guild.members.fetch(userId).catch(() => null);
-        if (!member) return res.status(404).json({ error: 'Member not found' });
-        await member.timeout(null, `Timeout removed via dashboard by ${req.user.username}`);
-        await logAuditEvent(req, guildId, 'untimeout', { targetUserId: userId });
-        res.json({ success: true });
-    } catch (error) {
-        console.error('Remove timeout error:', error);
-        res.status(500).json({ error: error.message || 'Internal server error' });
-    }
-});
-
-// ── Economy Admin ──────────────────────────────────────────────────────────
-
-router.get('/guild/:guildId/economy/stats', checkAuth, checkGuildAccess, async (req, res) => {
-    const { guildId } = req.params;
-    try {
-        const [topEarners, totalCoinsAgg, activeUsersCount, guildSettings] = await Promise.all([
-            User.aggregate([
-                { $match: { guildId } },
-                { $addFields: { total: { $add: ['$balance', '$bank'] } } },
-                { $sort: { total: -1 } },
-                { $limit: 10 },
-                { $project: { _id: 0, userId: 1, balance: 1, bank: 1, total: 1 } }
-            ]),
-            User.aggregate([{ $match: { guildId } }, { $group: { _id: null, total: { $sum: { $add: ['$balance', '$bank'] } } } }]),
-            User.countDocuments({ guildId, $or: [
-                { lastWork:  { $gte: new Date(Date.now() - 7 * 864e5) } },
-                { lastDaily: { $gte: new Date(Date.now() - 7 * 864e5) } },
-                { lastFish:  { $gte: new Date(Date.now() - 7 * 864e5) } },
-                { lastMine:  { $gte: new Date(Date.now() - 7 * 864e5) } },
-                { lastCrime: { $gte: new Date(Date.now() - 7 * 864e5) } },
-                { lastHeist: { $gte: new Date(Date.now() - 7 * 864e5) } },
-                { lastRob:   { $gte: new Date(Date.now() - 7 * 864e5) } }
-            ] }),
-            Guild.findOne({ guildId }).lean()
-        ]);
-
-        const commandUsage = guildSettings?.analytics?.commandUsage || [];
-        const econCommands = ['balance', 'daily', 'work', 'shop', 'rob', 'crime', 'duel', 'mine', 'fish', 'hunt', 'bank', 'pay', 'coinflip', 'roll', 'blackjack', 'casino'];
-        const commandFrequency = {};
-        for (const ev of commandUsage) {
-            if (econCommands.includes(ev.command)) {
-                commandFrequency[ev.command] = (commandFrequency[ev.command] || 0) + 1;
-            }
-        }
-
-        const ecoUserMap = {};
-        await Promise.all(topEarners.map(async u => {
-            try {
-                const user = await req.client.users.fetch(u.userId, { force: false });
-                ecoUserMap[u.userId] = { tag: user.tag, avatarUrl: user.displayAvatarURL({ size: 32, extension: 'webp' }) };
-            } catch { /* user not resolvable */ }
-        }));
-
-        res.json({
-            totalCoins: totalCoinsAgg[0]?.total || 0,
-            activeUsers: activeUsersCount,
-            topEarners: topEarners.map(u => ({
-                userId: u.userId,
-                userTag: ecoUserMap[u.userId]?.tag || null,
-                avatarUrl: ecoUserMap[u.userId]?.avatarUrl || null,
-                balance: u.balance, bank: u.bank,
-                total: (u.balance || 0) + (u.bank || 0)
-            })),
-            commandFrequency: Object.entries(commandFrequency).sort((a, b) => b[1] - a[1]).slice(0, 10).map(([cmd, count]) => ({ cmd, count }))
-        });
-    } catch (error) {
-        console.error('Economy stats error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.post('/guild/:guildId/economy/adjust', checkAuth, checkGuildAccess, checkCsrfOrigin, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { userId, action, amount } = req.body;
-
-    if (!userId || !isValidDiscordId(String(userId))) return res.status(400).json({ error: 'userId must be a valid Discord snowflake' });
-    if (!action || !['give', 'take', 'reset', 'freeze', 'unfreeze'].includes(action)) {
-        return res.status(400).json({ error: 'action must be give, take, reset, freeze, or unfreeze' });
-    }
-    if (['give', 'take'].includes(action)) {
-        const amt = Number(amount);
-        if (!Number.isFinite(amt) || amt <= 0 || !Number.isInteger(amt)) {
-            return res.status(400).json({ error: 'amount must be a positive integer for give/take' });
-        }
-    }
-
-    try {
-        const filter = { userId: String(userId), guildId };
-        let update;
-        if (action === 'give') {
-            update = { $inc: { balance: Number(amount) } };
-        } else if (action === 'take') {
-            // Use aggregation pipeline update to clamp balance at 0 atomically.
-            update = [{ $set: { balance: { $max: [0, { $subtract: ['$balance', Number(amount)] }] } } }];
-        } else if (action === 'reset') {
-            update = { $set: { balance: 0, bank: 0 } };
-        } else if (action === 'freeze') {
-            update = { $set: { economyFrozen: true } };
-        } else {
-            update = { $set: { economyFrozen: false } };
-        }
-
-        const user = await User.findOneAndUpdate(filter, update, { upsert: true, new: true, setDefaultsOnInsert: true });
-        await logAuditEvent(req, guildId, 'economy_adjust', { targetUserId: String(userId), action, amount: amount ?? null });
-        res.json({ success: true, balance: user.balance, bank: user.bank, economyFrozen: user.economyFrozen });
-    } catch (error) {
-        console.error('Economy adjust error:', error);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// ── Leveling admin endpoints ────────────────────────────────────────────────
-
-router.get('/guild/:guildId/leveling/leaderboard', checkAuth, checkGuildAccess, async (req, res) => {
-    const { guildId } = req.params;
-    const page = Math.max(1, parseInt(req.query.page, 10) || 1);
-    const limit = 25;
-    const skip = (page - 1) * limit;
-    try {
-        const [users, total] = await Promise.all([
-            User.find({ guildId, $or: [{ level: { $gt: 0 } }, { xp: { $gt: 0 } }] }).sort({ level: -1, xp: -1 }).skip(skip).limit(limit).select('userId level xp messages'),
-            User.countDocuments({ guildId, $or: [{ level: { $gt: 0 } }, { xp: { $gt: 0 } }] })
-        ]);
-        res.json({
-            entries: users.map((u, i) => ({ rank: skip + i + 1, userId: u.userId, level: u.level, xp: u.xp, messages: u.messages })),
-            total,
-            page,
-            pages: Math.ceil(total / limit)
-        });
-    } catch (err) {
-        console.error('Leveling leaderboard error:', err);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.post('/guild/:guildId/leveling/adjust', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { userId, action, amount } = req.body;
-    if (!userId || !isValidDiscordId(String(userId))) return res.status(400).json({ error: 'userId must be a valid Discord snowflake' });
-    if (!action || !['give', 'take', 'reset', 'set_level'].includes(action)) {
-        return res.status(400).json({ error: 'action must be give, take, reset, or set_level' });
-    }
-    if (['give', 'take', 'set_level'].includes(action)) {
-        const amt = Number(amount);
-        if (!Number.isFinite(amt) || !Number.isInteger(amt)) {
-            return res.status(400).json({ error: 'amount must be an integer' });
-        }
-        if (['give', 'take'].includes(action) && amt <= 0) {
-            return res.status(400).json({ error: 'amount must be positive for give/take' });
-        }
-        if (action === 'set_level' && amt < 0) {
-            return res.status(400).json({ error: 'level cannot be negative' });
-        }
-    }
-    try {
-        const filter = { userId: String(userId), guildId };
-        let update;
-        if (action === 'give') {
-            update = { $inc: { xp: Number(amount) } };
-        } else if (action === 'take') {
-            update = [{ $set: { xp: { $max: [0, { $subtract: ['$xp', Number(amount)] }] } } }];
-        } else if (action === 'reset') {
-            update = { $set: { xp: 0, level: 0 } };
-        } else {
-            update = { $set: { level: Number(amount) } };
-        }
-        const user = await User.findOneAndUpdate(filter, update, { upsert: true, new: true, setDefaultsOnInsert: true });
-        res.json({ success: true, level: user.level, xp: user.xp });
-    } catch (err) {
-        console.error('Leveling adjust error:', err);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-router.post('/guild/:guildId/leveling/xp-event', checkAuth, checkGuildAccess, checkWriteRateLimit, async (req, res) => {
-    const { guildId } = req.params;
-    const { multiplier, durationHours } = req.body;
-    const mult = Number(multiplier);
-    const hours = Number(durationHours);
-    if (!Number.isFinite(mult) || mult < 1.1 || mult > 10) return res.status(400).json({ error: 'multiplier must be between 1.1 and 10' });
-    if (!Number.isFinite(hours) || hours < 1 || hours > 168) return res.status(400).json({ error: 'durationHours must be between 1 and 168' });
-    try {
-        const existing = await Guild.findOne({ guildId }).select('leveling.xpBoostEvent').lean();
-        const existingEvent = existing?.leveling?.xpBoostEvent;
-        const isActive = existingEvent?.multiplier && existingEvent?.endTime && new Date(existingEvent.endTime).getTime() > Date.now();
-        const startTime = new Date();
-        const endTime = new Date(startTime.getTime() + hours * 3600 * 1000);
-        await Guild.findOneAndUpdate({ guildId }, {
-            $set: { 'leveling.xpBoostEvent.multiplier': mult, 'leveling.xpBoostEvent.startTime': startTime, 'leveling.xpBoostEvent.endTime': endTime }
-        }, { upsert: true });
-        res.json({ success: true, multiplier: mult, startTime, endTime, replacedActive: !!isActive });
-    } catch (err) {
-        console.error('XP event error:', err);
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
+router.use(require('./api/settings'));
+router.use(require('./api/stats'));
+router.use(require('./api/autorole'));
+router.use(require('./api/reactionRoles'));
+router.use(require('./api/rss'));
+router.use(require('./api/knowledgeBase'));
+router.use(require('./api/summaryJobs'));
+router.use(require('./api/dailyDigest'));
+router.use(require('./api/ai'));
+router.use(require('./api/members'));
+router.use(require('./api/achievements'));
+router.use(require('./api/itemImages'));
+router.use(require('./api/moderation'));
+router.use(require('./api/economy'));
+router.use(require('./api/leveling'));
 
 module.exports = router;
-module.exports.computeRetention = computeRetention;
-module.exports.validateEventLogUpdate = validateEventLogUpdate;
